remove file url page, directly on chat page after login
