+++ src/App.js
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import Navbar from "./components/Navbar";
-import FileUrlForm from "./components/FileUrlForm";
+// import FileUrlForm from "./components/FileUrlForm";
 import ChatPage from "./components/ChatPage";
 import Login from "./components/Login";
 import Register from "./components/Register";
@@ -30,14 +30,14 @@ function App() {
           <Route path="/forgot-password" element={<ForgotPassword />} />
           <Route path="/reset-password" element={<ResetPassword />} />
           <Route path="/register" element={<Register />} />
-          <Route
+          {/* <Route
             path="/"
             element={
               <ProtectedRoute>
                 <FileUrlForm setSubmittedData={setSubmittedData} />
               </ProtectedRoute>
             }
-          />
+          /> */}
           <Route
             path="/payment"
             element={
@@ -47,7 +47,7 @@ function App() {
             }
           />
           <Route
-            path="/chat"
+            path="/"
             element={
               <ProtectedRoute>
                 <ChatPage
