+++ src/components/Sidebar.js
@@ -1,7 +1,18 @@
-import React from "react";
-import { ListGroup } from "react-bootstrap";
+import React, { useState } from "react";
+import { ListGroup, Form, Button } from "react-bootstrap";
 
-function Sidebar({ files, urls, removeFile, removeUrl }) {
+function Sidebar({ prevfiles, urls = [], removeFile, removeUrl }) {
+  const [files, setFiles] = useState([]);
+  const [uploads, setUploads] = useState([]);
+  const handleFileChange = (event) => {
+    setUploads([...event.target.files]);
+  };
+
+  const handleSubmit = () => {
+    setFiles([...uploads]);
+    console.log("uploads: ", uploads);
+  };
+
   const marginStyle = { marginTop: "1.5cm" };
   const listStyle = {
     padding: "0.5rem",
@@ -9,25 +20,44 @@ function Sidebar({ files, urls, removeFile, removeUrl }) {
     wordWrap: "break-word",
     overflow: "hidden",
   };
+
   return (
     <div>
+      <Form style={marginStyle}>
+        <Form.Group className="mb-3">
+          <Form.Label>Add Files</Form.Label>
+          <Form.Control
+            type="file"
+            id="file"
+            accept=".txt,.pdf,.docx"
+            multiple
+            onChange={handleFileChange}
+          />
+        </Form.Group>
+        {uploads.length > 0 && (
+          <Button className="mt-3 w-100" onClick={handleSubmit}>
+            Submit
+          </Button>
+        )}
+      </Form>
       {files.length > 0 && <h2 style={marginStyle}>Files</h2>}
       <ListGroup>
         {files.map((file, index) => (
-          <ListGroup.Item
-            key={index}
-            className="d-flex justify-content-between align-items-center"
-            style={listStyle}
-          >
-            {file.name}
-            {/* <Button
+          <a href="/" key={index}>
+            <ListGroup.Item
+              className="d-flex justify-content-between align-items-center"
+              style={listStyle}
+            >
+              {file.name}
+              {/* <Button
               variant="danger"
               size="sm"
               onClick={() => removeFile(index)}
             >
               Remove
             </Button> */}
-          </ListGroup.Item>
+            </ListGroup.Item>
+          </a>
         ))}
       </ListGroup>
       {urls.length > 0 && <h2 style={marginStyle}>URLs</h2>}
