+++ src/index.css
@@ -9,6 +9,10 @@ code {
     monospace;
 }
 
+a {
+  text-decoration: none;
+}
+
 .gradient-custom-2 {
   /* fallback for old browsers */
   background: #fccb90;
